Rename adapter to compat
