+++ matchers/compat/parse.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package adapter
+package compat
 
 import (
 	"github.com/go-kit/log"
